Refactored files
Folders were restructured so that unit components could be processed without being lost between the pages displayed in full on the website
