+++ src/Components/Pages/Sales.jsx
@@ -1,0 +1,13 @@
+import { useUser } from "../../Context/UserContext";
+
+export default function Sales() {
+    const {mail} =useUser();
+    if(!mail) return null;
+    return(
+        <>
+        <form>
+            
+        </form>
+        </>
+    );
+}
